Makefile is changed to use compiler options from the common place Compiler warnings are fixed
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@486080 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/misc/src/main/native/accessors/shared/org_apache_harmony_misc_accessors_Malloc.c
@@ -110,7 +110,7 @@ JNIEXPORT jlong JNICALL Java_org_apache_harmony_misc_accessors_Malloc_memmove
 JNIEXPORT jlong JNICALL Java_org_apache_harmony_misc_accessors_Malloc_strncpy
   (JNIEnv *env, jclass self, jlong addr1, jlong addr2, jlong len)
 {
-    strncpy(jlong2addr(char, addr1), jlong2addr(const char, addr2), len);
+    strncpy(jlong2addr(char, addr1), jlong2addr(const char, addr2), (size_t)len);
     return addr1;
 }
